Found an eternal loop in the EntityMover
Removed Player.cs, this is not a game anymore, but a simulator.
EntityMover.cs has to be rewritten so it actually uses smart pathfinding
instead of walking in straight lines from roadpoint to roadpoint.
Dude.cs now gives more info about what's going on. This should be
changed so no text is given as args, so the logger handles that instead.